Changed Player Prefab & New Particle Texture
Changed Player Prefab to move the sprite renderer off the player. This is so that we do not accidently also rotate the camera when we rotate the sprite.

Camera rotation speed can be changed via sprite rather than player object

Additionally, added a new texture material to play when player gets hit.